Revert "Merge pull request #3 from mbreemhaar/feature/thiliumbot-features"
This reverts commit 9a34944bdc13378ded36442e3552a3f8a3def798, reversing
changes made to 57e1039597077ec5fb9c8984cc3ff79aca1b5dac.
